extract pupil identification flag service from pupil service
